% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ahccd_base_url}
\alias{ahccd_base_url}
\title{URLS for AHCCD collections}
\usage{
ahccd_base_url(collection = c("stations", "annual", "seasonal",
  "monthly"))
}
\arguments{
\item{collection}{character; return the URL for which collection?}
}
\value{
A character string containing the requested URL
}
\description{
Base URLs for AHCCD collections on GeoMet
}
\author{
Gavin L. Simpson
}
