% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ahccd_request}
\alias{ahccd_request}
\title{HTTP request from AHCCD}
\usage{
ahccd_request(url)
}
\arguments{
\item{url}{character; a URL querying GeoMet's OGC Features API}
}
\value{
The result of the HTTP requested, parsed from JSON.
}
\description{
Make a HTTP request for AHCCD records from GeoMet
}
\author{
Gavin L. Simpson
}
