% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_stations}
\alias{n_stations}
\title{Number of AHCCD stations}
\usage{
n_stations()
}
\value{
The number of stations as a length-1 integer
}
\description{
Returns the number of stations in the AHCCD with climate data
}
\author{
Gavin L. Simpson
}
