% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_list.R
\name{station_list}
\alias{station_list}
\title{List of AHCCD stations}
\usage{
station_list()
}
\value{
A tibble.
}
\description{
Queries the list of AHCCD stations from GeoMet
}
\author{
Gavin L. Simpson
}
